Removed warning under WIN64.

+++ copasi/compareExpressions/CNormalCall.cpp
@@ -1,10 +1,15 @@
 // Begin CVS Header
 //   $Source: /Volumes/Home/Users/shoops/cvs/copasi_dev/copasi/compareExpressions/CNormalCall.cpp,v $
-//   $Revision: 1.2 $
+//   $Revision: 1.2.4.1 $
 //   $Name:  $
-//   $Author: gauges $
-//   $Date: 2009/03/03 15:57:53 $
+//   $Author: shoops $
+//   $Date: 2011/01/10 17:00:27 $
 // End CVS Header
+
+// Copyright (C) 2010 by Pedro Mendes, Virginia Tech Intellectual
+// Properties, Inc., University of Heidelberg, and The University
+// of Manchester.
+// All rights reserved.
 
 // Copyright (C) 2008 by Pedro Mendes, Virginia Tech Intellectual
 // Properties, Inc., EML Research, gGmbH, University of Heidelberg,
@@ -33,7 +38,7 @@ CNormalCall::CNormalCall(const CNormalCall& src): CNormalBase(src)
 /**
  * Assignment operator
  */
-CNormalCall & CNormalCall::operator=(const CNormalCall& src)
+CNormalCall & CNormalCall::operator=(const CNormalCall & src)
 {
   if (&src != this)
     {
@@ -41,6 +46,7 @@ CNormalCall & CNormalCall::operator=(const CNormalCall& src)
       this->mType = src.mType;
       this->setFractions(src.mFractions);
     }
+
   return *this;
 }
 
@@ -48,50 +54,54 @@ CNormalCall & CNormalCall::operator=(const CNormalCall& src)
  * Smaller operator
  */
 bool CNormalCall::operator<(const CNormalCall& rhs) const
-  {
-    bool result = false;
-    if (this->mType < rhs.mType)
-      {
-        result = true;
-      }
-    else
-      {
-        if (this->mName < rhs.mName)
-          {
-            result = true;
-          }
-        else
-          {
-            if (this->mName == rhs.mName)
-              {
-                if (this->mFractions.size() < rhs.mFractions.size())
-                  {
-                    result = true;
-                  }
-                else
-                  {
-                    if (this->mFractions.size() == rhs.mFractions.size())
-                      {
-                        bool smaller = true;
-                        std::vector<CNormalFraction*>::const_iterator it = this->mFractions.begin(), endit = this->mFractions.end();
-                        std::vector<CNormalFraction*>::const_iterator it2 = rhs.mFractions.begin();
-                        while (it != endit && smaller == true)
-                          {
-                            smaller = ((*it2) < (*it));
-                            ++it;
-                            ++it2;
-                          }
-                        if (smaller)
-                          {
-                            result = true;
-                          }
-                      }
-                  }
-              }
-          }
-      }
-    return result;
-  }
+{
+  bool result = false;
+
+  if (this->mType < rhs.mType)
+    {
+      result = true;
+    }
+  else
+    {
+      if (this->mName < rhs.mName)
+        {
+          result = true;
+        }
+      else
+        {
+          if (this->mName == rhs.mName)
+            {
+              if (this->mFractions.size() < rhs.mFractions.size())
+                {
+                  result = true;
+                }
+              else
+                {
+                  if (this->mFractions.size() == rhs.mFractions.size())
+                    {
+                      bool smaller = true;
+                      std::vector<CNormalFraction*>::const_iterator it = this->mFractions.begin(), endit = this->mFractions.end();
+                      std::vector<CNormalFraction*>::const_iterator it2 = rhs.mFractions.begin();
+
+                      while (it != endit && smaller == true)
+                        {
+                          smaller = ((*it2) < (*it));
+                          ++it;
+                          ++it2;
+                        }
+
+                      if (smaller)
+                        {
+                          result = true;
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+  return result;
+}
 
 /**
  * Destructor
@@ -100,6 +110,7 @@ CNormalCall::~CNormalCall()
 {
   // delete all fractions
   std::vector<CNormalFraction*>::iterator it = this->mFractions.begin(), endit = this->mFractions.end();
+
   while (it != endit)
     {
       delete *it;
@@ -108,18 +119,18 @@ CNormalCall::~CNormalCall()
 }
 
 CNormalBase * CNormalCall::copy() const
-  {
-    return new CNormalCall(*this);
-  }
+{
+  return new CNormalCall(*this);
+}
 
 /**
  * Retrieve the number of summands of this sum.
  * @return int
  */
-int CNormalCall::getSize() const
-  {
-    return this->mFractions.size();
-  }
+size_t CNormalCall::getSize() const
+{
+  return this->mFractions.size();
+}
 
 /**
  * Add fraction to this sum.
@@ -136,49 +147,52 @@ bool CNormalCall::add(const CNormalFraction& fraction)
  * @return mFractions.
  */
 const std::vector<CNormalFraction*>& CNormalCall::getFractions() const
-  {
-    return this->mFractions;
-  }
+{
+  return this->mFractions;
+}
 
 /**
  * Examine equality of two sums.
  * @return bool.
  */
 bool CNormalCall::operator==(const CNormalCall & rhs) const
-  {
-    bool result = true;
-    if (this->mType == rhs.mType)
-      {
-        if (this->mName == rhs.mName)
-          {
-            if (this->mFractions.size() == rhs.mFractions.size())
-              {
-                std::vector<CNormalFraction*>::const_iterator it = this->mFractions.begin();
-                std::vector<CNormalFraction*>::const_iterator it2 = rhs.mFractions.begin();
-                std::vector<CNormalFraction*>::const_iterator endit = this->mFractions.end();
-                while (it != endit && result == true)
-                  {
-                    result = ((**it) == (**it2));
-                    ++it;
-                    ++it2;
-                  }
-              }
-            else
-              {
-                result = false;
-              }
-          }
-        else
-          {
-            result = false;
-          }
-      }
-    else
-      {
-        result = false;
-      }
-    return result;
-  }
+{
+  bool result = true;
+
+  if (this->mType == rhs.mType)
+    {
+      if (this->mName == rhs.mName)
+        {
+          if (this->mFractions.size() == rhs.mFractions.size())
+            {
+              std::vector<CNormalFraction*>::const_iterator it = this->mFractions.begin();
+              std::vector<CNormalFraction*>::const_iterator it2 = rhs.mFractions.begin();
+              std::vector<CNormalFraction*>::const_iterator endit = this->mFractions.end();
+
+              while (it != endit && result == true)
+                {
+                  result = ((**it) == (**it2));
+                  ++it;
+                  ++it2;
+                }
+            }
+          else
+            {
+              result = false;
+            }
+        }
+      else
+        {
+          result = false;
+        }
+    }
+  else
+    {
+      result = false;
+    }
+
+  return result;
+}
 
 /**
  * Sets the fractions of this product.
@@ -187,13 +201,16 @@ void CNormalCall::setFractions(const std::vector<CNormalFraction*>& set)
 {
   // delete all fractions
   std::vector<CNormalFraction*>::iterator it = this->mFractions.begin(), endit = this->mFractions.end();
+
   while (it != endit)
     {
       delete *it;
       ++it;
     }
+
   this->mFractions.clear();
   std::vector<CNormalFraction*>::const_iterator it2 = set.begin(), endit2 = set.end();
+
   while (it2 != endit2)
     {
       this->add(**it2);
@@ -202,32 +219,37 @@ void CNormalCall::setFractions(const std::vector<CNormalFraction*>& set)
 }
 
 std::string CNormalCall::toString() const
-  {
-    std::ostringstream os;
-    os << this->mName << "(";
-    std::vector<CNormalFraction*>::const_iterator it = this->mFractions.begin(), endit = this->mFractions.end();
-    while (it != endit)
-      {
-        os << (**it);
-        ++it;
-        if (it != endit)
-          {
-            os << ",";
-          }
-      }
-    os << ")";
-    return os.str();
-  }
+{
+  std::ostringstream os;
+  os << this->mName << "(";
+  std::vector<CNormalFraction*>::const_iterator it = this->mFractions.begin(), endit = this->mFractions.end();
+
+  while (it != endit)
+    {
+      os << (**it);
+      ++it;
+
+      if (it != endit)
+        {
+          os << ",";
+        }
+    }
+
+  os << ")";
+  return os.str();
+}
 
 bool CNormalCall::simplify()
 {
   // simplify all children
   std::vector<CNormalFraction*>::iterator it = this->mFractions.begin(), endit = this->mFractions.end();
+
   while (it != endit)
     {
       (*it)->simplify();
       ++it;
     }
+
   return true;
 }
 
@@ -238,9 +260,9 @@ std::ostream & operator<< (std::ostream &os, const CNormalCall & d)
 }
 
 const std::string& CNormalCall::getName() const
-  {
-    return this->mName;
-  }
+{
+  return this->mName;
+}
 
 void CNormalCall::setName(const std::string& name)
 {
@@ -248,9 +270,9 @@ void CNormalCall::setName(const std::string& name)
 }
 
 CNormalCall::Type CNormalCall::getType() const
-  {
-    return this->mType;
-  }
+{
+  return this->mType;
+}
 
 void CNormalCall::setType(CNormalCall::Type type)
 {
